Refactor general helper into own directory
